backend/requirements.txt
django==5.1.2
djangorestframework==3.15.2
mysqlclient==2.2.5
django-cors-headers==4.4.0
pymysql==1.1.1
Pillow==10.4.0